requirements.txt
hubspot-api-client
google-cloud-aiplatform[adk,agent_engines]>=1.111
uvicorn 
fastmcp
python-dotenv